refactor: limit breaking changes for next version
