+++ types/module/testutil/helpers.go
@@ -1,6 +1,12 @@
 package testutil
 
 import (
+	coretesting "cosmossdk.io/core/testing"
+	"cosmossdk.io/log"
+	"cosmossdk.io/store"
+	"cosmossdk.io/store/metrics"
+	storetypes "cosmossdk.io/store/types"
+
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/codec/testutil"
@@ -77,4 +83,17 @@ type TestTxBuilder struct {
 
 func (b *TestTxBuilder) SetExtensionOptions(extOpts ...*types.Any) {
 	b.ExtOptions = extOpts
+}
+
+// CreateMultiStore is a helper for setting up multiple stores for provided modules.
+func CreateMultiStore(keys map[string]*storetypes.KVStoreKey, logger log.Logger) storetypes.CommitMultiStore {
+	db := coretesting.NewMemDB()
+	cms := store.NewCommitMultiStore(db, logger, metrics.NewNoOpMetrics())
+
+	for key := range keys {
+		cms.MountStoreWithDB(keys[key], storetypes.StoreTypeIAVL, db)
+	}
+
+	_ = cms.LoadLatestVersion()
+	return cms
 }
